backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scipy
pydantic
sqlalchemy
httpx
python-dotenv
apscheduler
groq
